Move module `visualization` in subfolder `src/bed_bpp_env`
