+++ src/bed_bpp_env/visualization/__init__.py
@@ -6,7 +6,8 @@ import json
 import pathlib
 import utils
 
-import visualization.colors.generate_color_database
+from bed_bpp_env.visualization.colors import generate_color_database
+
 
 OUTPUTDIRECTORY = pathlib.Path.joinpath(utils.OUTPUTDIRECTORY, "vis")
 OUTPUTDIRECTORY.mkdir(parents=True, exist_ok=True)
@@ -18,11 +19,6 @@ if isinstance(usedData, str):
 if not (usedData is None):
     colorDBFilename = pathlib.Path(__file__).parent.resolve().joinpath(f"colors/colordb_{usedData.name}")
     if not (colorDBFilename.exists()):
-        visualization.colors.generate_color_database.generateColorDatabase(usedData)
+        generate_color_database.generateColorDatabase(usedData)
     with open(colorDBFilename) as file:
         COLOR_DATABASE = json.load(file, parse_int=False)
-
-
-from visualization.Visualization import Visualization
-from visualization.PalletizingEnvironmentVisualization import PalletizingEnvironmentVisualization
-from visualization.Video import Video
