+++ src/bed_bpp_env/visualization/colors/__init__.py
@@ -1,5 +1,4 @@
-'''
+"""
 This package automatically generates a color database for a given input file, i.e., order data.
 The colors are used for the visualization with VTK and Blender.
-'''
-from . import generate_color_database
+"""
