+++ src/bed_bpp_env/visualization/colors/generate_color_database.py
